.history/requirements_20250222134436.txt
absl-py==1.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.6.0
array-record==0.4.0
asn1crypto==0.24.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
blinker==1.4
blis==0.7.11
boto3==1.35.24
botocore==1.35.24
Brlapi==0.7.0
bs4==0.0.2
cached-property==1.5.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.3
comm==0.2.1
command-not-found==0.3
confection==0.1.4
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.8
datasets==3.0.0
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
dill==0.3.8
distlib==0.3.6
distro==1.5.0
distro-info==0.23+ubuntu1.1
dm-tree==0.1.8
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.3
etils==1.3.0
executing==2.0.1
filelock==3.12.0
findspark==2.0.1
flatbuffers==23.5.26
fonttools==4.53.1
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl#sha256=f6a4fa799b1daf2a635fe55a7de6b344379dc542d54d4c490113f0bc94bed4b8
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl#sha256=37e9f1f6a278a5138fabdabcc92cc559da917f9b24c76f0adf6758720d7eab10
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.4.0
gensim==4.3.2
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
h5py==3.10.0
html5lib==1.1
httplib2==0.14.0
huggingface-hub==0.22.0
idna==2.8
importlib-metadata==7.0.1
importlib_resources==6.1.2
install==1.3.5
ipykernel==6.29.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==3.2.0
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.13.1
keras-core==0.1.5
keras-cv==0.6.4
keras-nlp==0.6.1
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.0.1
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.6
louis==3.12.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mpi4py==3.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
mysql-connector-python==8.3.0
namex==0.0.7
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
paramiko==2.7.1
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==3.5.1
preshed==3.0.9
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.5.1
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydantic==2.6.4
pydantic_core==2.16.3
pyDOE==0.3.8
pyDOE3==1.0.4
pydot==2.0.0
Pygments==2.17.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.16.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.14.0
pytz==2024.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.2
regex==2023.12.25
reportlab==3.5.34
requests==2.32.3
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rich==13.7.0
rsa==4.9
s3transfer==0.10.2
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
sentence-transformers==2.6.1
sentencepiece==0.2.0
simplejson==3.16.0
six==1.14.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
system-service==0.3
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-text==2.13.0
termcolor==2.4.0
texttable==1.6.2
tf-keras==2.15.0
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
torch==2.2.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
transformers==4.39.1
triton==2.2.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.23.0
wadllib==1.3.3
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==3.0.1
wrapt==1.16.0
xkit==0.0.0
xxhash==3.5.0
yarl==1.11.1
zipp==3.17.0
zope.interface==4.7.1
